Encapsulate file structure access of diary entries
Class EntryFile is responsible for all file name operations
and file access of the diary entries.

Saving of a diary entry is done in the EditPane.leaveEvent
to reduce saving frequency.

Diary entries are saved only if they have actual content.
This prevents the creation of "empty" diary entry files.